Crate festival_api

Festival to WAV API

A simple REST api which takes a request body of the form:

{
    "word": "university",
    "lang": "en",
    "speed": 0.7,
    "fmt": "wav"
}

and returns a file which may be streamed or played for a user.

Detailed documentation on how to use the API can be found here.

Deployment

Docker

git clone https://github.com/JosiahBull/festival-api
cd ./festival-api
cp .example.env .env
nano .env #Update required configuration options for TLS encryption

nano ./config/general.toml #Update any general configuration options you wish to include.
nano ./config/langs.toml #Update any special languages you wish to include (ensure to modify backend.Dockerfile to install them).

docker volume create api-pgdata

chmod +x ./initalize_server.sh

docker-compose --env-file .env up #This takes a long time

Development

Fedora

To develop this api, you must have Rust installed.

This api also depends on Postgres, Sox, and Festival, and the festvox_cmu_us_aew lang pack.

Dependencies

    curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
    
    sudo dnf groupinstall @development-tools @development-libraries

    sudo dnf install festival sox libpq libpq-devel
    cargo install diesel_cli --no-default-features --features postgres
    cargo install rustfmt #for automatic code formatting
    cargo install clippy #collection of useful lints
    cargo install cargo-tarpaulin #test coverage

    #Install default language pack
    wget http://www.festvox.org/packed/festival/2.5/voices/festvox_cmu_us_aew_cg.tar.gz
    tar -xf festvox_cmu_us_aew_cg.tar.gz
    sudo mkdir -p /usr/share/festival/voices/us/
    sudo cp -r festival/lib/voices/us/cmu_us_aew_cg/ /usr/share/festival/voices/us/
    rm -rd ./festival
    rm festvox_cmu_us_aew_cg.tar.gz

Development

docker run --name fest-db -p 5432:5432 -e POSTGRES_PASSWORD=postgres -d postgres -N 500
export DATABASE_URL=postgres://postgres:postgres@localhost/fest_api

diesel setup

export JWT_SECRET=<Your_Token_Here>
cargo test -- --test-threads 4
cargo run

Contributing

Code Guidelines

Please write tests if we have good test coverage we can avoid any bugs down the line.

Outside of this we use standard Rust formatting for code. This will be enforced through use of clippy and rustfmt.

Commit Guidelines

In all commits, please try to follow the convention for commits.

Ideally aim to push every commit you make, rather than accumulating a large number of commits before pushing, this helps to keep everyone on the same codebase when collaborating.

The exception for this is that you should not commit non-compiling code to the main branch. Open a new branch and commit to that instead.

Use of Pull Requests

Outside of exceptional cases, please always push commits to a new branch and then generate a pull request with your new feature. Automated actions will attempt to validate that your code does not break anything existing, but this doesn’t guarantee your code works. Please write tests!

Modules

common

Common functions used in endpoints. Varied between db interactions and general processing.

macros

This module contains ease-of-use macros for the Festival-Api.

models

Various objects, including database objects, for the api.

Structs

DbConn

Functions

convert

Expects a phrase package, attempts to convert it to a sound file to be returned to the user. Requires an authenticate user account to access. This endpoint also features strict rate limiting as generating .wav files is very resource intensive.

create

Attempt to create a new user account

docs

Returns the OAS docs for this api in an easily downloadable file.

index

The base url of the program. This is just a catch-all for those who stumble across the api without knowing what it does.

login

Attempts to login a student with provided credentials.

main