Module festival_api::cache

This module handles caching for the api. A configurable number of words, or a max total size of files should be kept. The data flow is as follows req -> check if item in cache -> item in cache, increment popularity score, return item -> not in cache, generate item fresh, update popularity score -> if space in cache, save item. -> if no space in cache, check if our score is greater (10pts?) than the lowest popularity score -> if greater, bump that item, update the lowest popularity score number, cache this item -> else, do nothing.

The API for this cache is designed to be flexible and reusable, so it makes extensive use of generics. This means that to use it expect to have to implement functions to pass into the api.

Structs

Cache

A cache for storing frequently used data. Will automatically attempt to cache popular items over time, decaching less popular items as required.

Info

A struct which wraps cacheable data for the cache. This is useful as we can store information about the stored item, which can be used when making decisions about whether to cache or not.

Constants

BYTES_IN_MB

The number of bytes in a mb.

Traits

Cachable

A type indicating that this return type may be cached properly in the api. These methods are required to save and load from the disk.