Trait festival_api::cache::Cachable

trait Cachable<U> {
    fn load_underlying<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<U, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn size_on_disk<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<usize, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn save_on_disk<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn remove_from_disk<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

A type indicating that this return type may be cached properly in the api. These methods are required to save and load from the disk.

Required methods

fn load_underlying<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<U, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Load the underlying file from the disk, or generate it fresh if it’s not there.

fn size_on_disk<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<usize, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get the total size this object is taking up on the disk. Should fail if the file has not been saved to the disk.

fn save_on_disk<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Save this file to the disk, so that it is cached for future use. May require regeneration of the file.

fn remove_from_disk<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Remove the underlying file from the disk.