Module festival_api::common

Enums

SearchItem

Functions

compare_hashed_strings

A function which checks whether the first string can be hashed into the second string. Returns a boolean true if they are the same, and false otherwise.

find_user_in_db

Attempt to find a user in the database, returns None if the user is unable to be found. Note that the provided name is assumed unique. If multiple results exist, the first will be returned.

generate_random_alphanumeric

Generate a randomised alphanumeric (base 62) string of a requested length.

get_time_since

Get the time (in seconds) since a chrono datetime. Returns a duration which can be negative if the time is in the future.

hash_string_with_salt

Hash a string with a random salt to be stored in the database. Utilizes the argon2id algorithm Followed best practices as laid out here: https://cheatsheetseries.owasp.org/cheatsheets/Password_Storage_Cheat_Sheet.html

load_recent_requests

Load a users most recent requests, limited based on

update_user_last_seen

Attempts to find and then update a user with a new timestamp.