Function festival_api::common::find_user_in_db

pub async fn find_user_in_db(
    conn: &DbConn, 
    name: SearchItem
) -> Result<Option<User>, Response>

Attempt to find a user in the database, returns None if the user is unable to be found. Note that the provided name is assumed unique. If multiple results exist, the first will be returned.