Function festival_api::convert

pub async fn convert(
    phrase_package: Json<PhrasePackage>, 
    generator: &Flite, 
    converter: &Converter, 
    cfg: &Config, 
    cache: Cache
) -> Result<Response, Response>

Expects a phrase package, attempts to convert it to a sound file to be returned to the user. Requires an authenticate user account to access. This endpoint also features strict rate limiting as generating .wav files is very resource intensive.