Macro festival_api::macros::failure

macro_rules! failure {
    () => { ... };
    ($arg : tt) => { ... };
    ($($arg : tt) *) => { ... };
}

A macro to shorthand the rejection from an endpoint due to a server error. /// Should be used when you want a quick 500 response to the user. Note that a string must be provided (or used like format!()).

Examples

    fn main() {
        #[get("/")]
        fn index() -> Response {
            failure!("The server had a critical error processing your request!");
        }
    }

/// ```rust fn main() { #[get(“/”)] fn index() -> Response { let reason: String = “it caught fire!”; failure!(“The server failed to process your request becuase {}”, reason); } }

If you need a more detailed failure response other htan 500 + a message
please construct the response manually.