Macro festival_api::macros::load_env

macro_rules! load_env {
    () => { ... };
    ($arg : tt) => { ... };
    ($($arg : tt) *) => { ... };
}

A macro to load configuration from the environment.

Attempts to load from multiple sources falling back in this order:

Load from environment

Load from ./config/general.toml

panic!

This macro recommend for use in conjunction with lazy static, as these variables like to be loaded/parsed at runtime, not at compile-time.

Example

    lazy_static! {
        static ref NUMBER_SHOES: usize = load_env!("NUMBER_SHOES");
    }

    fn main() {
        lazy_static::initialize(&NUMBER_SHOES);
        println!("The number of shoes is {}", *NUMBER_SHOES);
    }

A variety of types are supported for implicit conversion, look here for a dedicated list of these types.

Internally this macro relies on toml::value::Value.try_into() for type conversion.