Module festival_api::models

Structs

Claims

The claims held by the JWT used for authentication

GenerationRequest

Language

NewGenerationRequest

PhrasePackage

A phrase package which the user is requesting a .mp3 for

User

Represents a user of this api.

UserCredentials

User credentials, to be used when logging in or creating a new account

Functions

_impl_insertable_for_newgenerationrequest

_impl_insertable_for_usercredentials

_impl_queryable_by_name_for_generationrequest

_impl_queryable_by_name_for_user

_impl_queryable_for_generationrequest

_impl_queryable_for_user