Crate festival_api

Modules

cache

This module handles caching for the api. A configurable number of words, or a max total size of files should be kept. The data flow is as follows req -> check if item in cache -> item in cache, increment popularity score, return item -> not in cache, generate item fresh, update popularity score -> if space in cache, save item. -> if no space in cache, check if our score is greater (10pts?) than the lowest popularity score -> if greater, bump that item, update the lowest popularity score number, cache this item -> else, do nothing.

common

macros

This module contains ease-of-use macros for the Festival-Api.

models

response

schema

Structs

ALLOWED_FORMATS

The list of supported file-formats, note that wav is the preferred format due to lower cpu usage.

API_NAME

The name of the api which is sent with certain requests.

CACHE_PATH

The path to the cache for storing .wav files.

DbConn

JWT_EXPIRY_TIME_HOURS

The number of hours that a JWT may be used before expiring and forcing the user to revalidate.

JWT_SECRET

The secret used for fast-hashing JWT’s for validation.

MAX_REQUESTS_ACC_THRESHOLD

The maximum requests that an account can make in a given time period established by MAX_REQUESTS_TIME_PERIOD_MINUTES

MAX_REQUESTS_TIME_PERIOD_MINUTES

The time period for timing out users who make too many requests.

SPEED_MAX_VAL

The maximum speed at which a phrase can be read.

SPEED_MIN_VAL

The lowerest speed at which a phrase can be read.

SUPPORTED_LANGS

A list of supported speech languages by this api.

TEMP_PATH

The path where temporary files are stored, and should be deleted from on a crash.

WORD_LENGTH_LIMIT

The maximum length of a phrase that the api will process.

Functions

convert

Expects a phrase package, attempts to convert it to a .mp3 to be returned to the user. Requires authentication to access.

create

Attempt to create a new user account

docs

index

login

Attempt to login a student

main

rocket